# [JSON-API] key_hash field to speed up fetchByKey queries