--- a/ledger-service/db-backend/src/main/scala/com/digitalasset/http/dbbackend/Queries.scala
+++ b/ledger-service/db-backend/src/main/scala/com/digitalasset/http/dbbackend/Queries.scala
@@ -26,6 +26,7 @@ import cats.instances.list._
 import cats.Applicative
 import cats.syntax.applicative._
 import cats.syntax.functor._
+import com.daml.lf.crypto.Hash
 import doobie.free.connection
 
 sealed abstract class Queries(tablePrefix: String) {
@@ -67,6 +68,7 @@ sealed abstract class Queries(tablePrefix: String) {
         (contract_id $contractIdType NOT NULL CONSTRAINT ${tablePrefixFr}contract_k PRIMARY KEY
         ,tpid $bigIntType NOT NULL REFERENCES $templateIdTableName (tpid)
         ,${jsonColumn(sql"key")}
+        ,key_hash $keyHashColumn
         ,${jsonColumn(contractColumnName)}
         $contractsTableSignatoriesObservers
         ,agreement_text $agreementTextType
@@ -103,6 +105,7 @@ sealed abstract class Queries(tablePrefix: String) {
   protected[this] def maxListSize: Option[Int]
 
   protected[this] def jsonColumn(name: Fragment): Fragment
+  protected[this] def keyHashColumn: Fragment
 
   private[this] val createTemplateIdsTable = CreateTable(
     templateIdTableNameRaw,
@@ -330,12 +333,13 @@ sealed abstract class Queries(tablePrefix: String) {
     ): Query0[DBContract[Mark0, JsValue, JsValue, Vector[String]]] = {
       val q = query(tpid, queriesCondition)
       q.query[
-        (String, Mark0, Key, JsValue, SigsObs, SigsObs, Agreement)
-      ].map { case (cid, tpid, rawKey, payload, signatories, observers, rawAgreement) =>
+        (String, Mark0, Key, String, JsValue, SigsObs, SigsObs, Agreement)
+      ].map { case (cid, tpid, rawKey, key_hash, payload, signatories, observers, rawAgreement) =>
         DBContract(
           contractId = cid,
           templateId = tpid,
           key = key(rawKey),
+          keyHash = key_hash,
           payload = payload,
           signatories = sigsObs(signatories),
           observers = sigsObs(observers),
@@ -366,14 +370,14 @@ sealed abstract class Queries(tablePrefix: String) {
   private[http] final def fetchByKey(
       parties: OneAnd[Set, String],
       tpid: SurrogateTpId,
-      key: JsValue,
+      key: Hash,
   )(implicit
       log: LogHandler,
       ipol: SqlInterpol,
   ): ConnectionIO[Option[DBContract[Unit, JsValue, JsValue, Vector[String]]]] =
     selectContracts(parties, tpid, keyEquality(key)).option
 
-  private[http] def keyEquality(key: JsValue): Fragment
+  private[http] def keyEquality(key: Hash): Fragment = sql"key_hash = ${key.toHexString.toString}"
 
   private[http] def equalAtContractPath(path: JsonPath, literal: JsValue): Fragment
 
@@ -406,6 +410,7 @@ object Queries {
       contractId: String,
       templateId: TpId,
       key: CK,
+      keyHash: String,
       payload: PL,
       signatories: Prt,
       observers: Prt,
@@ -614,6 +619,7 @@ private final class PostgresQueries(tablePrefix: String) extends Queries(tablePr
   protected[this] override def agreementTextType = sql"TEXT NOT NULL"
 
   protected[this] override def jsonColumn(name: Fragment) = name ++ sql" JSONB NOT NULL"
+  protected[this] override def keyHashColumn = textType
 
   protected[this] override val maxListSize = None
 
@@ -629,8 +635,13 @@ private final class PostgresQueries(tablePrefix: String) extends Queries(tablePr
       CREATE INDEX $contractsTableIndexName ON $contractTableName (tpid)
     """)
 
+  private[this] val contractKeyHashIndexName = Fragment.const0(s"${tablePrefix}ckey_hash_idx")
+  private[this] val contractKeyHashIndex = CreateIndex(
+    sql"""CREATE INDEX $contractKeyHashIndexName ON $contractTableName (key_hash)"""
+  )
+
   protected[this] override def initDatabaseDdls =
-    super.initDatabaseDdls ++ Seq(indexContractsTable, indexContractsKeys)
+    super.initDatabaseDdls ++ Seq(indexContractsTable, indexContractsKeys, contractKeyHashIndex)
 
   protected[http] override def version()(implicit log: LogHandler): ConnectionIO[Option[Int]] = {
     for {
@@ -661,7 +672,7 @@ private final class PostgresQueries(tablePrefix: String) extends Queries(tablePr
     Update[DBContract[SurrogateTpId, JsValue, JsValue, Array[String]]](
       s"""
         INSERT INTO $contractTableNameRaw
-        VALUES (?, ?, ?::jsonb, ?::jsonb, ?, ?, ?)
+        VALUES (?, ?, ?::jsonb, ?, ?::jsonb, ?, ?, ?)
         ON CONFLICT (contract_id) DO NOTHING
       """
     ).updateMany(dbcs)
@@ -682,7 +693,7 @@ private final class PostgresQueries(tablePrefix: String) extends Queries(tablePr
       trackMatchIndices,
       tpidSelector = fr"tpid",
       query = (tpid, unionPred) =>
-        sql"""SELECT contract_id, $tpid tpid, key, payload, signatories, observers, agreement_text
+        sql"""SELECT contract_id, $tpid tpid, key, key_hash, payload, signatories, observers, agreement_text
               FROM $contractTableName AS c
               WHERE (signatories && $partyVector::text[] OR observers && $partyVector::text[])
                     AND ($unionPred)""",
@@ -699,9 +710,6 @@ private final class PostgresQueries(tablePrefix: String) extends Queries(tablePr
         path.elems.map(_.fold(k => sql"->${k: String}", (_: _0.type) => sql"->0")),
       )
     )
-
-  private[http] override def keyEquality(key: JsValue): Fragment =
-    sql"key = $key::jsonb"
 
   private[http] override def equalAtContractPath(path: JsonPath, literal: JsValue) =
     fragmentContractPath(path) ++ sql" = ${literal}::jsonb"
@@ -758,6 +766,8 @@ private final class OracleQueries(tablePrefix: String) extends Queries(tablePref
   protected[this] override def jsonColumn(name: Fragment) =
     sql"$name CLOB NOT NULL CONSTRAINT ${tablePrefixFr}ensure_json_$name CHECK ($name IS JSON)"
 
+  protected[this] override def keyHashColumn = sql"NVARCHAR2(64)"
+
   // See http://www.dba-oracle.com/t_ora_01795_maximum_number_of_expressions_in_a_list_is_1000.htm
   protected[this] override def maxListSize = Some(1000)
 
@@ -782,8 +792,13 @@ private final class OracleQueries(tablePrefix: String) extends Queries(tablePref
     sql"""CREATE INDEX $stakeholdersIndexName ON $contractStakeholdersViewName (tpid, stakeholder)"""
   )
 
+  private[this] val contractKeyHashIndexName = Fragment.const0(s"${tablePrefix}ckeyHash_idx")
+  private[this] def contractKeyHashIndex = CreateIndex(
+    sql"""CREATE INDEX $contractKeyHashIndexName ON $contractTableName (key_hash)"""
+  )
+
   protected[this] override def initDatabaseDdls =
-    super.initDatabaseDdls ++ Seq(stakeholdersView, stakeholdersIndex)
+    super.initDatabaseDdls ++ Seq(stakeholdersView, stakeholdersIndex, contractKeyHashIndex)
 
   protected[http] override def version()(implicit log: LogHandler): ConnectionIO[Option[Int]] = {
     import cats.implicits._
@@ -811,11 +826,11 @@ private final class OracleQueries(tablePrefix: String) extends Queries(tablePref
       dbcs: F[DBContract[SurrogateTpId, DBContractKey, JsValue, Array[String]]]
   )(implicit log: LogHandler, ipol: SqlInterpol): ConnectionIO[Int] = {
     import spray.json.DefaultJsonProtocol._
-    Update[DBContract[SurrogateTpId, JsValue, JsValue, JsValue]](
+    Update[DBContract[SurrogateTpId, DBContractKey, JsValue, JsValue]](
       s"""
         INSERT /*+ ignore_row_on_dupkey_index($contractTableNameRaw(contract_id)) */
-        INTO $contractTableNameRaw (contract_id, tpid, key, payload, signatories, observers, agreement_text)
-        VALUES (?, ?, ?, ?, ?, ?, ?)
+        INTO $contractTableNameRaw (contract_id, tpid, key, key_hash, payload, signatories, observers, agreement_text)
+        VALUES (?, ?, ?, ?, ?, ?, ?, ?)
       """
     ).updateMany(
       dbcs.map(_.mapKeyPayloadParties(identity, identity, _.toJson))
@@ -842,7 +857,7 @@ private final class OracleQueries(tablePrefix: String) extends Queries(tablePref
           sql"""contract_id, template_id, key, payload,
                 signatories, observers, agreement_text"""
         val dupQ =
-          sql"""SELECT c.contract_id contract_id, $tpid template_id, key, payload,
+          sql"""SELECT c.contract_id contract_id, $tpid template_id, key, key_hash, payload,
                        signatories, observers, agreement_text ${rownum getOrElse fr""}
                 FROM $contractTableName c
                      JOIN $contractStakeholdersViewName cst ON (c.contract_id = cst.contract_id)
@@ -857,11 +872,6 @@ private final class OracleQueries(tablePrefix: String) extends Queries(tablePref
       },
       agreement = (_: Option[String]) getOrElse "",
     )
-
-  private[http] override def keyEquality(key: JsValue): Fragment = {
-    import spray.json.DefaultJsonProtocol.JsValueFormat
-    sql"JSON_EQUAL(key, ${toDBContractKey(key)})"
-  }
 
   private[this] def pathSteps(path: JsonPath): Cord =
     path.elems.foldMap(_.fold(k => (".\"": Cord) ++ k :- '"', (_: _0.type) => "[0]"))