--- a/ledger-service/http-json/src/bench/scala/com/daml/http/dbbackend/ContractDaoBenchmark.scala
+++ b/ledger-service/http-json/src/bench/scala/com/daml/http/dbbackend/ContractDaoBenchmark.scala
@@ -56,6 +56,7 @@ abstract class ContractDaoBenchmark extends OracleAround {
     contractId = s"#$id",
     templateId = tpid,
     key = JsNull,
+    keyHash = "",
     payload = payload,
     signatories = Seq(signatory),
     observers = Seq.empty,