--- a/ledger-service/http-json/src/main/scala/com/digitalasset/http/ContractsFetch.scala
+++ b/ledger-service/http-json/src/main/scala/com/digitalasset/http/ContractsFetch.scala
@@ -178,6 +178,7 @@ private class ContractsFetch(
   ): Exception \/ PreInsertContract = {
     import scalaz.syntax.traverse._
     import scalaz.std.option._
+    import com.daml.lf.crypto.Hash
     for {
       ac <- domain.ActiveContract fromLedgerApi ce leftMap (de =>
         new IllegalArgumentException(s"contract ${ce.contractId}: ${de.shows}"): Exception,
@@ -188,6 +189,13 @@ private class ContractsFetch(
       contractId = ac.contractId.unwrap,
       templateId = ac.templateId,
       key = lfKey.cata(lfValueToDbJsValue, JsNull),
+      keyHash = lfKey.cata(
+        Hash
+          .assertHashContractKey(TemplateId.toLedgerApiValue(ac.templateId), _)
+          .toHexString
+          .toString,
+        "",
+      ),
       payload = lfValueToDbJsValue(lfArg),
       signatories = ac.signatories,
       observers = ac.observers,