Addition of a key_hash field to speed up fetchByKey queries

Based on gatling perf tests

baseline
```
> Number of requests                                  1000 (OK=1000   KO=-     )
> Min. response time                                   487 (OK=487    KO=-     )
> Max. response time                                  4631 (OK=4631   KO=-     )
> Mean response time                                  2462 (OK=2462   KO=-     )
> Std. deviation                                       787 (OK=787    KO=-     )
> response time 90th percentile                       2507 (OK=2507   KO=-     )
> response time 95th percentile                       3002 (OK=3002   KO=-     )
> response time 99th percentile                       3637 (OK=3637   KO=-     )
> response time 99.9th percentile                     4126 (OK=4126   KO=-     )
> Mean requests/second                               3.906 (OK=3.906  KO=-     )
---- Response time distribution ------------------------------------------------
> t < 800 ms                                             7 (  1%)
> 800 ms < t < 1200 ms                                  67 (  7%)
> t > 1200 ms                                          926 ( 93%)
> failed                                                 0 (  -%)
```

query using hashed key field
```
> Number of requests                                  1000 (OK=1000   KO=-     )
> Min. response time                                   206 (OK=206    KO=-     )
> Max. response time                                  2893 (OK=2893   KO=-     )
> Mean response time                                  1074 (OK=1074   KO=-     )
> Std. deviation                                       348 (OK=348    KO=-     )
> response time 90th percentile                       1100 (OK=1100   KO=-     )
> response time 95th percentile                       1330 (OK=1330   KO=-     )
> response time 99th percentile                       1577 (OK=1577   KO=-     )
> response time 99.9th percentile                     1832 (OK=1832   KO=-     )
> Mean requests/second                                8.85 (OK=8.85   KO=-     )
---- Response time distribution ------------------------------------------------
> t < 800 ms                                           217 ( 22%)
> 800 ms < t < 1200 ms                                 422 ( 42%)
> t > 1200 ms                                          361 ( 36%)
> failed                                                 0 (  -%)
```

changelog_begin
- Update schema version for http-json-api query store with new key_hash field
- Improved performance for fetchByKey query which now uses key_hash field
changelog_end

### Pull Request Checklist

- [x] Read and understand the [contribution guidelines](https://github.com/digital-asset/daml/blob/main/CONTRIBUTING.md)
- [x] Include appropriate tests
- [x] Set a descriptive title and thorough description
- [x] Add a reference to the [issue this PR will solve](https://github.com/digital-asset/daml/issues), if appropriate
- [x] Include changelog additions in one or more commit message bodies between the `CHANGELOG_BEGIN` and `CHANGELOG_END` tags
- [ ] Normal production system change, include purpose of change in description

NOTE: CI is not automatically run on non-members pull-requests for security
reasons. The reviewer will have to comment with `/AzurePipelines run` to
trigger the build.