@stefanobaghino-da 2021-08-20
In Oracle test results:

```
enterprise-19.12.0-preloaded-20210817-24-6f020d6: Pulling from digitalasset/oracle
Digest: sha256:7af81f12931bc6673578b5aca651b871c998332e6013f5a738184f62ceece8c0
Status: Image is up to date for digitalasset/oracle:enterprise-19.12.0-preloaded-20210817-24-6f020d6
docker.io/digitalasset/oracle:enterprise-19.12.0-preloaded-20210817-24-6f020d6
Error: No such container: oracle
5d3d2508d6474966e584e0ea7f76beb839e9554d7e0f043ff94f837b98f8fb8f
Could not connect to Oracle, trying again...
Could not connect to Oracle, trying again...
Could not connect to Oracle, trying again...
```

🤔